+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 3.0.1-dev-1.0 (2022-09-16)
+
+### BREAKING CHANGES
+
+* old/unused code and webpack removed
+* Android: Black Screen fixed
+* callbacks changed to Promises
+
+
 <a name="3.0.1"></a>
 ## [3.0.1](https://github.com/bitpay/cordova-plugin-qrscanner/compare/2.6.0...3.0.1) (2019-03-29)